Add cloudwatch to loggly lambda task.
This task is not deployed via serverless, because
it also required specific KMS encrpytion setup and
an associated IAM role, which I think would not
be possible via serverless, or at least would be
hard.

All new lambda tasks can have there output logged
to loggly by adding thier CWL group to the event
source for the cwlogs-to-loggly lambda function.